# frontend/src/types/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Squant",
    "description": "Personal Quantitative Trading System for Cryptocurrency",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/exchange-accounts/{account_id}/test": {
      "post": {
        "tags": [
          "Exchange Accounts"
        ],
        "summary": "Test Exchange Connection",
        "description": "Test connection to exchange using stored credentials.\n\nAttempts to fetch account balance to verify API credentials are valid.\n\nArgs:\n    account_id: Account UUID.\n    session: Database session.\n\nReturns:\n    Connection test result with success status and balance count.\n\nRaises:\n    HTTPException: 404 if account not found.",
        "operationId": "test_exchange_connection_api_v1_exchange_accounts__account_id__test_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ConnectionTestResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/backtest": {
      "post": {
        "tags": [
          "Backtest"
        ],
        "summary": "Run Backtest",
        "description": "Create and run a backtest synchronously.\n\nThis endpoint creates a backtest run and executes it immediately,\nreturning when the backtest completes.\n\nArgs:\n    request: Backtest configuration.\n    session: Database session.\n\nReturns:\n    Backtest run with results.\n\nRaises:\n    HTTPException: 400 if insufficient data, 404 if strategy not found.",
        "operationId": "run_backtest_api_v1_backtest_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunBacktestRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BacktestRunResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Backtest"
        ],
        "summary": "List Backtests",
        "description": "List backtest runs with optional filters and pagination.\n\nArgs:\n    strategy_id: Optional strategy ID to filter by.\n    status: Optional status filter (pending, running, completed, error).\n    page: Page number (1-indexed).\n    page_size: Items per page.\n    session: Database session.\n\nReturns:\n    Paginated list of backtest runs.",
        "operationId": "list_backtests_api_v1_backtest_get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Strategy ID (optional)",
              "title": "Strategy Id"
            },
            "description": "Strategy ID (optional)"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Status filter: pending, running, completed, error",
              "title": "Status"
            },
            "description": "Status filter: pending, running, completed, error"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Page size"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PaginatedData_BacktestListItem__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/async": {
      "post": {
        "tags": [
          "Backtest"
        ],
        "summary": "Create Backtest",
        "description": "Create a backtest run without executing it.\n\nUse this to create a backtest that can be executed later.\nCall POST /backtest/{run_id}/run to execute.\n\nArgs:\n    request: Backtest configuration.\n    session: Database session.\n\nReturns:\n    Created backtest run (status: pending).\n\nRaises:\n    HTTPException: 404 if strategy not found.",
        "operationId": "create_backtest_api_v1_backtest_async_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BacktestRunResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/{run_id}/run": {
      "post": {
        "tags": [
          "Backtest"
        ],
        "summary": "Execute Backtest",
        "description": "Execute a pending backtest.\n\nArgs:\n    run_id: Backtest run ID.\n    session: Database session.\n\nReturns:\n    Updated backtest run with results.\n\nRaises:\n    HTTPException: 400 if execution fails or data incomplete, 404 if not found.",
        "operationId": "execute_backtest_api_v1_backtest__run_id__run_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BacktestRunResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/{run_id}/cancel": {
      "post": {
        "tags": [
          "Backtest"
        ],
        "summary": "Cancel Backtest",
        "description": "Cancel a running backtest (TRD-008#3).\n\nThis endpoint requests cancellation of a running backtest.\nThe backtest will stop at the next bar iteration.\n\nArgs:\n    run_id: Backtest run ID.\n    session: Database session.\n\nReturns:\n    Current backtest run state.\n\nRaises:\n    HTTPException: 400 if not cancellable, 404 if not found.",
        "operationId": "cancel_backtest_api_v1_backtest__run_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BacktestRunResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/data/check": {
      "post": {
        "tags": [
          "Backtest"
        ],
        "summary": "Check Data Availability",
        "description": "Check historical data availability for backtesting.\n\nArgs:\n    request: Data availability check request.\n    session: Database session.\n\nReturns:\n    Data availability information.",
        "operationId": "check_data_availability_api_v1_backtest_data_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DataAvailabilityResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/data/symbols": {
      "get": {
        "tags": [
          "Backtest"
        ],
        "summary": "List Available Symbols",
        "description": "List symbols with available historical data.\n\nArgs:\n    exchange: Optional exchange filter.\n    timeframe: Optional timeframe filter.\n    session: Database session.\n\nReturns:\n    List of available symbols with data info.",
        "operationId": "list_available_symbols_api_v1_backtest_data_symbols_get",
        "parameters": [
          {
            "name": "exchange",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by exchange",
              "title": "Exchange"
            },
            "description": "Filter by exchange"
          },
          {
            "name": "timeframe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by timeframe",
              "title": "Timeframe"
            },
            "description": "Filter by timeframe"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_AvailableSymbolResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/{run_id}": {
      "get": {
        "tags": [
          "Backtest"
        ],
        "summary": "Get Backtest",
        "description": "Get a backtest run by ID.\n\nArgs:\n    run_id: Backtest run ID.\n    session: Database session.\n\nReturns:\n    Backtest run details.\n\nRaises:\n    HTTPException: 404 if not found.",
        "operationId": "get_backtest_api_v1_backtest__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BacktestRunResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Backtest"
        ],
        "summary": "Delete Backtest",
        "description": "Delete a backtest run.\n\nArgs:\n    run_id: Backtest run ID.\n    session: Database session.\n\nReturns:\n    Success response.\n\nRaises:\n    HTTPException: 404 if not found.",
        "operationId": "delete_backtest_api_v1_backtest__run_id__delete",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/backtest/{run_id}/detail": {
      "get": {
        "tags": [
          "Backtest"
        ],
        "summary": "Get Backtest Detail",
        "description": "Get backtest details including equity curve.\n\nArgs:\n    run_id: Backtest run ID.\n    session: Database session.\n\nReturns:\n    Detailed backtest response with equity curve.\n\nRaises:\n    HTTPException: 404 if not found.",
        "operationId": "get_backtest_detail_api_v1_backtest__run_id__detail_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_BacktestDetailResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_BacktestRunResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/BacktestRunResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[BacktestRunResponse]"
      },
...
      "ApiResponse_ConnectionTestResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/ConnectionTestResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[ConnectionTestResponse]"
      },
      "ApiResponse_DataAvailabilityResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/DataAvailabilityResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[DataAvailabilityResponse]"
      },
...
      "ApiResponse_PaginatedData_BacktestListItem__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "$ref": "#/components/schemas/PaginatedData_BacktestListItem_"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[PaginatedData[BacktestListItem]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}